feat(devnet): allow devs to mint eth
